+++ docassemble_webapp/docassemble/webapp/file_access.py
@@ -199,7 +199,7 @@ def get_info_from_file_reference(file_reference, **kwargs):
         elif len(parts) == 2:
             result['package'] = parts[0]
         result['fullpath'] = docassemble.base.functions.static_filename_path(file_reference)
-    #logmessage("path is " + str(result['fullpath']))
+    # sys.stderr.write("path is " + str(result['fullpath']) + "\n")
     if result['fullpath'] is not None: #os.path.isfile(result['fullpath'])
         if not has_info:
             result['filename'] = os.path.basename(result['fullpath'])
@@ -217,13 +217,13 @@ def get_info_from_file_reference(file_reference, **kwargs):
                 result['fullpath'] = result['path'] + '.' + result['extension']
                 ext_type, result['mimetype'] = get_ext_and_mimetype(result['fullpath'])
             else:
-                logmessage("Did not find file " + result['path'] + '.' + convert[result['extension']])
+                sys.stderr.write("Did not find file " + result['path'] + '.' + convert[result['extension']] + "\n")
                 return dict()
         #logmessage("Full path is " + result['fullpath'])
         if os.path.isfile(result['fullpath']) and not has_info:
             add_info_about_file(result['fullpath'], result['path'], result)
     else:
-        logmessage("File reference " + str(file_reference) + " DID NOT EXIST.")
+        sys.stderr.write("File reference " + str(file_reference) + " DID NOT EXIST.\n")
     return(result)
 
 def add_info_about_file(filename, basename, result):